daisy UI 적용

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -15,15 +15,15 @@
         <tbody>
           <tr>
             <th>번호</th>
-            <td>${article.id}</td>
+           <td><div class="badge badge-primary">${article.id}</div></td>
           </tr>
           <tr>
             <th>작성날짜</th>
-            <td>${article.regDate.substring(2, 16)}</td>
+            <td>${article.regDateForPrint}</td>
           </tr>
           <tr>
             <th>수정날짜</th>
-            <td>${article.updateDate.substring(2, 16)}</td>
+            <td>${article.updateDateForPrint}</td>
           </tr>
           <tr>
             <th>작성자</th>
@@ -42,12 +42,13 @@
     </div>
 
     <div class="btns">
-      <button class="btn-text-link" type="button" onclick="history.back();">뒤로가기</button>
-      <a class="btn-text-link" href="../article/modify?id=${article.id}">게시물 수정</a>
-      <c:if test="${article.extra__actorCanDelete}">
-        <a class="btn-text-link" onclick="if ( confirm('정말 삭제하시겠습니까?') == false ) return false;"
-          href="../article/doDelete?id=${article.id}">게시물 삭제</a>
-      </c:if>
+       <button class="btn btn-link" type="button" onclick="history.back();">뒤로가기</button>
+       <c:if test="${article.extra__actorCanModify}">
+         <a class="btn btn-link" href="../article/modify?id=${article.id}">게시물 수정</a>
+       </c:if>
+       <c:if test="${article.extra__actorCanDelete}">
+         <a class="btn btn-link" onclick="if ( confirm('정말 삭제하시겠습니까?') == false ) return false;" href="../article/doDelete?id=${article.id}">게시물 삭제</a>
+       </c:if>
     </div>
   </div>
 </section>
